% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bbox_cidr}
\alias{bbox_cidr}
\title{Bounding box from CIDR blocks}
\usage{
bbox_cidr(cidr, addr_space_bits_per_image = 32L,
  addr_space_bits_per_pixel = 8L)
}
\arguments{
\item{cidr}{character vector of dotted-decimal/digit CIDRs}
}
\value{
list of bounding box extents for each CIDR - \code{cidr},
\code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}
}
\description{
Returns a data frame of bounding boxes for a given CIDR within the
Hilbert-curve plane.
}
\examples{
bbox_cidr("30.0.0.0/8")
}
