% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-hilbertv4.R
\docType{data}
\name{CoordHilbertV4}
\alias{CoordHilbertV4}
\alias{coord_hilbert_v4}
\title{ggplot2 `Coord` for Hilbert IPv4 heatmaps}
\format{An object of class \code{CoordHilbertV4} (inherits from \code{CoordCartesian}, \code{Coord}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
CoordHilbertV4

coord_hilbert_v4(xlim = NULL, ylim = NULL, bpp = c(8, 10, 12, 14, 16),
  expand = FALSE)
}
\arguments{
\item{xlim}{Limits for the x and y axes.}

\item{ylim}{Limits for the x and y axes.}

\item{bpp}{specifies the number of address space bits assigned to each pixel
in the output image. By default each pixel represents a /24 network,
which corresponds to 8 host bits (i.e., 256 hosts). Valid values are
8, 10, 12, 14, and 16 (for now).}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}
}
\description{
ggplot2 `Coord` for Hilbert IPv4 heatmaps
}
\examples{
# only doing this to generate some random values to show
ip_df <- data.frame(ip=sample(400000000, 1000000, replace=TRUE))

ggplot(ip_df, aes(ip=ip)) +
  stat_hilbert_v4(bpp=16) +
  coord_hilbert_v4(bpp=16) +
  viridis::scale_fill_viridis(name="IPv4 count per pixel", trans="log2") +
  theme_hilbert_v4()
}
\keyword{datasets}
