% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-hilbertv4.R
\docType{data}
\name{StatHilbertV4}
\alias{StatHilbertV4}
\alias{stat_hilbert_v4}
\title{ggplot2 `Stat` for Hilbert IPv4 heatmaps}
\format{An object of class \code{StatHilbertV4} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 7.}
\usage{
StatHilbertV4

stat_hilbert_v4(mapping = NULL, data = NULL, geom = "raster", bpi = 32,
  bpp = c(8, 10, 12, 14, 16), position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{bpi}{bits per image (very likely changing to a CIDR notation)}

\item{bpp}{specifies the number of address space bits assigned to each pixel
in the output image. By default each pixel represents a /24 network,
which corresponds to 8 host bits (i.e., 256 hosts). Valid values are
8, 10, 12, 14, and 16 (for now).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{...}{other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
ggplot2 `Stat` for Hilbert IPv4 heatmaps
}
\examples{
# only doing this to generate some random values to show
ip_df <- data.frame(ip=sample(400000000, 1000000, replace=TRUE))

ggplot(ip_df, aes(ip=ip)) +
  stat_hilbert_v4(bpp=16) +
  coord_hilbert_v4(bpp=16) +
  viridis::scale_fill_viridis(name="IPv4 count per pixel", trans="log2") +
  theme_hilbert_v4()
}
\keyword{datasets}
