% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-hilbertv4.R
\name{theme_hilbert_v4}
\alias{theme_hilbert_v4}
\title{"Map" theme for Hilbert IPv4 maps}
\usage{
theme_hilbert_v4(fill = "black")
}
\arguments{
\item{fill}{background fill color (defaults to "black")}
}
\description{
"Map" theme for Hilbert IPv4 maps
}
\examples{
# only doing this to generate some random values to show
ip_df <- data.frame(ip=sample(400000000, 1000000, replace=TRUE))

ggplot(ip_df, aes(ip=ip)) +
  stat_hilbert_v4(bpp=16) +
  coord_hilbert_v4(bpp=16) +
  viridis::scale_fill_viridis(name="IPv4 count per pixel", trans="log2") +
  theme_hilbert_v4()
}
